target-roas calculation for internal
