+++ src/internal/business/dto/calculate-target-roas.dto.ts
@@ -1,0 +1,10 @@
+import { IsNotEmpty, IsNumber, IsPositive } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+
+export class CalculateTargetRoasDto {
+  @ApiProperty()
+  @IsNumber()
+  @IsPositive()
+  @IsNotEmpty()
+  budget: number;
+}
