+++ src/utils/platform.ts
@@ -1,0 +1,5 @@
+export enum Platform {
+  Facebook = 'facebook',
+  GoogleSearch = 'googleSearch',
+  Instagram = 'instagram',
+}
